# Move extensions out of tree.