--- a/Makefile
+++ b/Makefile
@@ -2,9 +2,7 @@ BIN_DIR=_output/bin
 RELEASE_VER=0.2
 
 kube-arbitrator: init
-	go build -o ${BIN_DIR}/kar-scheduler ./cmd/kar-scheduler/
-	go build -o ${BIN_DIR}/kar-controllers ./cmd/kar-controllers/
-	go build -o ${BIN_DIR}/karcli ./cmd/karcli
+	go build -o ${BIN_DIR}/kube-batchd ./cmd/kube-batchd/
 
 verify: generate-code
 	hack/verify-gofmt.sh
@@ -17,15 +15,11 @@ init:
 generate-code:
 	go build -o ${BIN_DIR}/deepcopy-gen ./cmd/deepcopy-gen/
 	${BIN_DIR}/deepcopy-gen -i ./pkg/apis/scheduling/v1alpha1/ -O zz_generated.deepcopy
-	${BIN_DIR}/deepcopy-gen -i ./pkg/apis/extensions/v1alpha1/ -O zz_generated.deepcopy
 
 images: kube-arbitrator
-	cp ./_output/bin/kar-scheduler ./deployment/
-	cp ./_output/bin/kar-controllers ./deployment/
-	cp ./_output/bin/karcli ./deployment/
-	docker build ./deployment/ -f ./deployment/Dockerfile.sched -t kubearbitrator/kar-scheduler:${RELEASE_VER}
-	docker build ./deployment/ -f ./deployment/Dockerfile.ctrl -t kubearbitrator/kar-controllers:${RELEASE_VER}
-	rm -f ./deployment/kar*
+	cp ./_output/bin/kube-batchd ./deployment/
+	docker build ./deployment/ -f ./deployment/Dockerfile.sched -t kubearbitrator/kube-batchd:${RELEASE_VER}
+	rm -f ./deployment/kube-batchd
 
 run-test:
 	hack/make-rules/test.sh $(WHAT) $(TESTS)