--- a/config/crds/scheduling_v1alpha1_podgroup.yaml
+++ b/config/crds/scheduling_v1alpha1_podgroup.yaml
@@ -1,9 +1,9 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: podgroups.scheduling.arbitrator.incubator.k8s.io
+  name: podgroups.scheduling.incubator.k8s.io
 spec:
-  group: scheduling.arbitrator.incubator.k8s.io
+  group: scheduling.incubator.k8s.io
   names:
     kind: PodGroup
     plural: podgroups