--- a/config/crds/scheduling_v1alpha1_queue.yaml
+++ b/config/crds/scheduling_v1alpha1_queue.yaml
@@ -1,9 +1,9 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: queues.scheduling.arbitrator.incubator.k8s.io
+  name: queues.scheduling.incubator.k8s.io
 spec:
-  group: scheduling.arbitrator.incubator.k8s.io
+  group: scheduling.incubator.k8s.io
   names:
     kind: Queue
     plural: queues