--- a/hack/.golint_failures
+++ b/hack/.golint_failures
@@ -1,13 +1,7 @@
-cmd/kar-controllers/app
-cmd/kar-controllers/app/options
-cmd/kar-scheduler/app
-cmd/kar-scheduler/app/options
-pkg/apis/extensions/v1alpha1
+cmd/kube-batchd/app
+cmd/kube-batchd/app/options
 pkg/apis/scheduling/v1alpha1
 pkg/apis/utils
-pkg/controller/job
-pkg/controller/queue
-pkg/karcli/job
 pkg/scheduler
 pkg/scheduler/actions/allocate
 pkg/scheduler/actions/preempt