--- a/hack/run-e2e.sh
+++ b/hack/run-e2e.sh
@@ -1,26 +1,27 @@
 #!/bin/bash
 
 export PATH="${HOME}/.kubeadm-dind-cluster:${PATH}"
+export KA_BIN=_output/bin
+export LOG_LEVEL=3
 
 # start k8s dind cluster
 ./hack/dind-cluster-v1.11.sh up
 
 kubectl create -f config/crds/scheduling_v1alpha1_podgroup.yaml
-kubectl create -f config/crds/extensions_v1alpha1_job.yaml
 kubectl create -f config/crds/scheduling_v1alpha1_queue.yaml
 
 # start kube-arbitrator
-nohup _output/bin/kar-controllers --kubeconfig ${HOME}/.kube/config --logtostderr --v 3 > controller.log 2>&1 &
-nohup _output/bin/kar-scheduler --kubeconfig ${HOME}/.kube/config --logtostderr --v 3 > scheduler.log 2>&1 &
+nohup ${KA_BIN}/kube-batchd --kubeconfig ${HOME}/.kube/config --logtostderr --v ${LOG_LEVEL} > scheduler.log 2>&1 &
 
 # clean up
 function cleanup {
-    killall -9 kar-controllers kar-scheduler
+    killall -9 kube-batchd
     ./hack/dind-cluster-v1.11.sh down
 
-    echo "=================================> Controller Logs <================================="
-    cat controller.log
-    echo "=================================> Scheduler Logs <================================="
+    echo "===================================================================================="
+    echo "=============================>>>>> Scheduler Logs <<<<<============================="
+    echo "===================================================================================="
+
     cat scheduler.log
 }