--- a/pkg/apis/scheduling/v1alpha1/register.go
+++ b/pkg/apis/scheduling/v1alpha1/register.go
@@ -27,11 +27,16 @@ var (
 	AddToScheme   = SchemeBuilder.AddToScheme
 )
 
-// GroupName is the group name used in this package.
-const GroupName = "scheduling.arbitrator.incubator.k8s.io"
+const (
+	// GroupName is the group name used in this package.
+	GroupName = "scheduling.incubator.k8s.io"
+
+	// GroupVersion is the version of scheduling group
+	GroupVersion = "v1alpha1"
+)
 
 // SchemeGroupVersion is the group version used to register these objects.
-var SchemeGroupVersion = schema.GroupVersion{Group: GroupName, Version: "v1alpha1"}
+var SchemeGroupVersion = schema.GroupVersion{Group: GroupName, Version: GroupVersion}
 
 // Resource takes an unqualified resource and returns a Group-qualified GroupResource.
 func Resource(resource string) schema.GroupResource {