… cmd/kar-controllers/app/options/options.go
… cmd/kar-controllers/app/server.go
… cmd/kar-scheduler/main.go
… cmd/karcli/karcli.go
… config/crds/extensions_v1alpha1_job.yaml
… pkg/apis/extensions/v1alpha1/doc.go
… pkg/apis/extensions/v1alpha1/job.go
… pkg/apis/extensions/v1alpha1/register.go
… pkg/apis/extensions/v1alpha1/zz_generated.deepcopy.go
… pkg/client/clientset/versioned/typed/extensions/v1alpha1/doc.go
… pkg/client/clientset/versioned/typed/extensions/v1alpha1/extensions_client.go
… pkg/client/clientset/versioned/typed/extensions/v1alpha1/fake/doc.go
… pkg/client/clientset/versioned/typed/extensions/v1alpha1/fake/fake_extensions_client.go
… pkg/client/clientset/versioned/typed/extensions/v1alpha1/fake/fake_job.go
… pkg/client/clientset/versioned/typed/extensions/v1alpha1/generated_expansion.go
… pkg/client/clientset/versioned/typed/extensions/v1alpha1/job.go
… pkg/client/informers/externalversions/extensions/interface.go
… pkg/client/informers/externalversions/extensions/v1alpha1/interface.go
… pkg/client/informers/externalversions/extensions/v1alpha1/job.go
… pkg/client/listers/extensions/v1alpha1/expansion_generated.go
… pkg/client/listers/extensions/v1alpha1/job.go
… pkg/controller/job/job_controller.go
… pkg/controller/job/utils.go
… pkg/controller/queue/queue_controller.go
… pkg/controller/queue/utils.go
… pkg/karcli/job/common.go
… pkg/karcli/job/list.go
… pkg/karcli/job/run.go
… pkg/karcli/job/util.go
… test/util.go