@TravisBuddy 2018-09-04
## Travis tests have failed

Hey @k82cn,
Please read the following log in order to understand the failure reason.
It'll be awesome if you fix what's wrong and commit the changes.


### 1st Build


<details>
  <summary>
    <strong>
     make e2e
    </strong>
  </summary>

```
mkdir -p _output/bin
go build -o _output/bin/kube-batchd ./cmd/kube-batchd/
hack/run-e2e.sh
* Making sure DIND image is up to date 
v1.11: Pulling from mirantis/kubeadm-dind-cluster




















Digest: sha256:ecc1e36f98124e3dae1112b35c82a2ace59851d4b232941f6d0f66bc27264645
Status: Downloaded newer image for mirantis/kubeadm-dind-cluster:v1.11
/home/travis/.kubeadm-dind-cluster/kubectl-v1.11.0: OK
* Starting DIND container: kube-master
* Running kubeadm: init --config /etc/kubeadm.conf --ignore-preflight-errors=all
Initializing machine ID from random generator.
Synchronizing state of docker.service with SysV service script with /lib/systemd/systemd-sysv-install.
Executing: /lib/systemd/systemd-sysv-install enable docker
Loaded image: mirantis/hypokube:base

real	0m3.626s
user	0m0.204s
sys	0m0.180s

Step 1/2 : FROM mirantis/hypokube:base
 ---> 8d0cea4f3fb8
Step 2/2 : COPY hyperkube /hyperkube
 ---> 3c3defbff1d3
Removing intermediate container 6087f0f0af3c
Successfully built 3c3defbff1d3
Created symlink /etc/systemd/system/multi-user.target.wants/kubelet.service → /lib/systemd/system/kubelet.service.
I0904 07:23:59.932413     495 feature_gate.go:230] feature gates: &{map[]}
[init] using Kubernetes version: v1.11.0
[preflight] running pre-flight checks
	[WARNING FileContent--proc-sys-net-bridge-bridge-nf-call-iptables]: /proc/sys/net/bridge/bridge-nf-call-iptables does not exist
	[WARNING FileExisting-crictl]: crictl not found in system path
I0904 07:23:59.977414     495 kernel_validator.go:81] Validating kernel version
I0904 07:23:59.977729     495 kernel_validator.go:96] Validating kernel config
[preflight/images] Pulling images required for setting up a Kubernetes cluster
[preflight/images] This might take a minute or two, depending on the speed of your internet connection
[preflight/images] You can also perform this action in beforehand using 'kubeadm config images pull'
[kubelet] Writing kubelet environment file with flags to file "/var/lib/kubelet/kubeadm-flags.env"
[kubelet] Writing kubelet configuration to file "/var/lib/kubelet/config.yaml"
[preflight] Activating the kubelet service
[certificates] Generated ca certificate and key.
[certificates] Generated apiserver certificate and key.
[certificates] apiserver serving cert is signed for DNS names [kube-master kubernetes kubernetes.default kubernetes.default.svc kubernetes.default.svc.cluster.local] and IPs [10.96.0.1 172.18.0.2]
[certificates] Generated apiserver-kubelet-client certificate and key.
[certificates] Generated sa key and public key.
[certificates] Generated front-proxy-ca certificate and key.
[certificates] Generated front-proxy-client certificate and key.
[certificates] Generated etcd/ca certificate and key.
[certificates] Generated etcd/server certificate and key.
[certificates] etcd/server serving cert is signed for DNS names [kube-master localhost] and IPs [127.0.0.1 ::1]
[certificates] Generated etcd/peer certificate and key.
[certificates] etcd/peer serving cert is signed for DNS names [kube-master localhost] and IPs [172.18.0.2 127.0.0.1 ::1]
[certificates] Generated etcd/healthcheck-client certificate and key.
[certificates] Generated apiserver-etcd-client certificate and key.
[certificates] valid certificates and keys now exist in "/etc/kubernetes/pki"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/admin.conf"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/kubelet.conf"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/controller-manager.conf"
[kubeconfig] Wrote KubeConfig file to disk: "/etc/kubernetes/scheduler.conf"
[controlplane] wrote Static Pod manifest for component kube-apiserver to "/etc/kubernetes/manifests/kube-apiserver.yaml"
[controlplane] wrote Static Pod manifest for component kube-controller-manager to "/etc/kubernetes/manifests/kube-controller-manager.yaml"
[controlplane] wrote Static Pod manifest for component kube-scheduler to "/etc/kubernetes/manifests/kube-scheduler.yaml"
[etcd] Wrote Static Pod manifest for a local etcd instance to "/etc/kubernetes/manifests/etcd.yaml"
[init] waiting for the kubelet to boot up the control plane as Static Pods from directory "/etc/kubernetes/manifests" 
[init] this might take a minute or longer if the control plane images have to be pulled
[apiclient] All control plane components are healthy after 41.501883 seconds
[uploadconfig] storing the configuration used in ConfigMap "kubeadm-config" in the "kube-system" Namespace
[kubelet] Creating a ConfigMap "kubelet-config-1.11" in namespace kube-system with the configuration for the kubelets in the cluster
[markmaster] Marking the node kube-master as master by adding the label "node-role.kubernetes.io/master=''"
[markmaster] Marking the node kube-master as master by adding the taints [node-role.kubernetes.io/master:NoSchedule]
[patchnode] Uploading the CRI Socket information "/var/run/dockershim.sock" to the Node API object "kube-master" as an annotation
[bootstraptoken] using token: npjyjs.v4ahzj3modi5s6n4
[bootstraptoken] configured RBAC rules to allow Node Bootstrap tokens to post CSRs in order for nodes to get long term certificate credentials
[bootstraptoken] configured RBAC rules to allow the csrapprover controller automatically approve CSRs from a Node Bootstrap Token
[bootstraptoken] configured RBAC rules to allow certificate rotation for all node client certificates in the cluster
[bootstraptoken] creating the "cluster-info" ConfigMap in the "kube-public" namespace
[addons] Applied essential addon: kube-dns
[addons] Applied essential addon: kube-proxy

Your Kubernetes master has initialized successfully!

To start using your cluster, you need to run the following as a regular user:

  mkdir -p $HOME/.kube
  sudo cp -i /etc/kubernetes/admin.conf $HOME/.kube/config
  sudo chown $(id -u):$(id -g) $HOME/.kube/config

You should now deploy a pod network to the cluster.
Run "kubectl apply -f [podnetwork].yaml" with one of the options listed at:
  https://kubernetes.io/docs/concepts/cluster-administration/addons/

You can now join any number of machines by running the following on each node
as root:

  kubeadm join 172.18.0.2:6443 --token npjyjs.v4ahzj3modi5s6n4 --discovery-token-ca-cert-hash sha256:7b6f84c5f8dc42cb500b63cc95cff1f2e8b8d5f20d2d2207491622c9e232995b


real	0m58.021s
user	0m4.628s
sys	0m0.164s
6c6593866c79
04fd6d18a87a
bb2394695a5d
b853911915ad
12d04273c680
2c13028013d1
e73457e3910e
9bca4db8efc5
Warning: kubectl apply should be used on resource created by either kubectl create --save-config or kubectl apply
configmap/kube-proxy configured
Warning: kubectl apply should be used on resource created by either kubectl create --save-config or kubectl apply
daemonset.extensions/kube-proxy configured
No resources found
* Setting cluster config 
Cluster "dind" set.
Context "dind" created.
Switched to context "dind".
* Starting node container: 1
* Starting DIND container: kube-node-1
* Node container started: 1
* Starting node container: 2
* Starting DIND container: kube-node-2
* Node container started: 2
* Joining node: 1
* Joining node: 2
* Running kubeadm: join --ignore-preflight-errors=all 172.18.0.2:6443 --token npjyjs.v4ahzj3modi5s6n4 --discovery-token-ca-cert-hash sha256:7b6f84c5f8dc42cb500b63cc95cff1f2e8b8d5f20d2d2207491622c9e232995b
* Running kubeadm: join --ignore-preflight-errors=all 172.18.0.2:6443 --token npjyjs.v4ahzj3modi5s6n4 --discovery-token-ca-cert-hash sha256:7b6f84c5f8dc42cb500b63cc95cff1f2e8b8d5f20d2d2207491622c9e232995b
Initializing machine ID from random generator.
Initializing machine ID from random generator.
Warning: Stopping docker.service, but it can still be activated by:
  docker.socket
Synchronizing state of docker.service with SysV service script with /lib/systemd/systemd-sysv-install.
Executing: /lib/systemd/systemd-sysv-install enable docker
Synchronizing state of docker.service with SysV service script with /lib/systemd/systemd-sysv-install.
Executing: /lib/systemd/systemd-sysv-install enable docker
Loaded image: mirantis/hypokube:base
Loaded image: mirantis/hypokube:base

real	0m5.908s
user	0m0.264s
sys	0m0.156s

real	0m5.918s
user	0m0.232s
sys	0m0.180s

Step 1/2 : FROM mirantis/hypokube:base
 ---> 8d0cea4f3fb8
Step 2/2 : COPY hyperkube /hyperkube

Step 1/2 : FROM mirantis/hypokube:base
 ---> 8d0cea4f3fb8
Step 2/2 : COPY hyperkube /hyperkube
 ---> 9a77d0c84473
 ---> 6ebf5b551f39
Removing intermediate container 8471777cca5a
Successfully built 9a77d0c84473
Removing intermediate container 3a36e96a36b0
Successfully built 6ebf5b551f39
Created symlink /etc/systemd/system/multi-user.target.wants/kubelet.service → /lib/systemd/system/kubelet.service.
Created symlink /etc/systemd/system/multi-user.target.wants/kubelet.service → /lib/systemd/system/kubelet.service.
[preflight] running pre-flight checks
[preflight] running pre-flight checks
	[WARNING RequiredIPVSKernelModulesAvailable]: the IPVS proxier will not be used, because the following required kernel modules are not loaded: [ip_vs_wrr ip_vs_sh ip_vs ip_vs_rr] or no builtin kernel ipvs support: map[ip_vs_rr:{} ip_vs_wrr:{} ip_vs_sh:{} nf_conntrack_ipv4:{} ip_vs:{}]
you can solve this problem with following methods:
 1. Run 'modprobe -- ' to load missing kernel modules;
2. Provide the missing builtin kernel ipvs support

	[WARNING RequiredIPVSKernelModulesAvailable]: the IPVS proxier will not be used, because the following required kernel modules are not loaded: [ip_vs ip_vs_rr ip_vs_wrr ip_vs_sh] or no builtin kernel ipvs support: map[ip_vs:{} ip_vs_rr:{} ip_vs_wrr:{} ip_vs_sh:{} nf_conntrack_ipv4:{}]
you can solve this problem with following methods:
 1. Run 'modprobe -- ' to load missing kernel modules;
2. Provide the missing builtin kernel ipvs support

	[WARNING FileContent--proc-sys-net-bridge-bridge-nf-call-iptables]: /proc/sys/net/bridge/bridge-nf-call-iptables does not exist
	[WARNING FileExisting-crictl]: crictl not found in system path
	[WARNING FileContent--proc-sys-net-bridge-bridge-nf-call-iptables]: /proc/sys/net/bridge/bridge-nf-call-iptables does not exist
	[WARNING FileExisting-crictl]: crictl not found in system path
I0904 07:26:03.616367     474 kernel_validator.go:81] Validating kernel version
I0904 07:26:03.616628     474 kernel_validator.go:96] Validating kernel config
I0904 07:26:03.619695     475 kernel_validator.go:81] Validating kernel version
I0904 07:26:03.619889     475 kernel_validator.go:96] Validating kernel config
[discovery] Trying to connect to API Server "172.18.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://172.18.0.2:6443"
[discovery] Trying to connect to API Server "172.18.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://172.18.0.2:6443"
[discovery] Failed to connect to API Server "172.18.0.2:6443": token id "npjyjs" is invalid for this cluster or it has expired. Use "kubeadm token create" on the master node to creating a new valid token
[discovery] Failed to connect to API Server "172.18.0.2:6443": token id "npjyjs" is invalid for this cluster or it has expired. Use "kubeadm token create" on the master node to creating a new valid token
[discovery] Trying to connect to API Server "172.18.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://172.18.0.2:6443"
[discovery] Trying to connect to API Server "172.18.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://172.18.0.2:6443"
[discovery] Failed to connect to API Server "172.18.0.2:6443": token id "npjyjs" is invalid for this cluster or it has expired. Use "kubeadm token create" on the master node to creating a new valid token
[discovery] Failed to connect to API Server "172.18.0.2:6443": token id "npjyjs" is invalid for this cluster or it has expired. Use "kubeadm token create" on the master node to creating a new valid token
[discovery] Trying to connect to API Server "172.18.0.2:6443"
[discovery] Trying to connect to API Server "172.18.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://172.18.0.2:6443"
[discovery] Created cluster-info discovery client, requesting info from "https://172.18.0.2:6443"
[discovery] Requesting info from "https://172.18.0.2:6443" again to validate TLS against the pinned public key
[discovery] Requesting info from "https://172.18.0.2:6443" again to validate TLS against the pinned public key
[discovery] Cluster info signature and contents are valid and TLS certificate validates against pinned roots, will use API Server "172.18.0.2:6443"
[discovery] Successfully established connection with API Server "172.18.0.2:6443"
[discovery] Cluster info signature and contents are valid and TLS certificate validates against pinned roots, will use API Server "172.18.0.2:6443"
[discovery] Successfully established connection with API Server "172.18.0.2:6443"
[kubelet] Downloading configuration for the kubelet from the "kubelet-config-1.11" ConfigMap in the kube-system namespace
[kubelet] Writing kubelet configuration to file "/var/lib/kubelet/config.yaml"
[kubelet] Downloading configuration for the kubelet from the "kubelet-config-1.11" ConfigMap in the kube-system namespace
[kubelet] Writing kubelet configuration to file "/var/lib/kubelet/config.yaml"
[kubelet] Writing kubelet environment file with flags to file "/var/lib/kubelet/kubeadm-flags.env"
[preflight] Activating the kubelet service
[kubelet] Writing kubelet environment file with flags to file "/var/lib/kubelet/kubeadm-flags.env"
[preflight] Activating the kubelet service
[tlsbootstrap] Waiting for the kubelet to perform the TLS Bootstrap...
[tlsbootstrap] Waiting for the kubelet to perform the TLS Bootstrap...
[patchnode] Uploading the CRI Socket information "/var/run/dockershim.sock" to the Node API object "kube-node-2" as an annotation
[patchnode] Uploading the CRI Socket information "/var/run/dockershim.sock" to the Node API object "kube-node-1" as an annotation

This node has joined the cluster:
* Certificate signing request was sent to master and a response
  was received.
* The Kubelet was informed of the new secure connection details.

Run 'kubectl get nodes' on the master to see this node join the cluster.

real	0m18.919s
user	0m0.616s
sys	0m0.080s
* Node joined: 1

This node has joined the cluster:
* Certificate signing request was sent to master and a response
  was received.
* The Kubelet was informed of the new secure connection details.

Run 'kubectl get nodes' on the master to see this node join the cluster.

real	0m18.908s
user	0m0.636s
sys	0m0.072s
* Node joined: 2
Creating static routes for bridge plugin
* Deploying k8s dashboard 
deployment.extensions/kubernetes-dashboard created
service/kubernetes-dashboard created
clusterrolebinding.rbac.authorization.k8s.io/add-on-cluster-admin created
* Patching kube-dns deployment to make it start faster 
Warning: kubectl apply should be used on resource created by either kubectl create --save-config or kubectl apply
deployment.extensions/kube-dns configured
* Taking snapshot of the cluster 
deployment.extensions/kube-dns scaled
deployment.extensions/kubernetes-dashboard scaled
WARNING: controller manager glitch: dns & dashboard won't stop; pods may 'blink' for some time after restore
tar: var/lib/kubelet/device-plugins/kubelet.sock: socket ignored
tar: var/lib/kubelet/device-plugins/kubelet.sock: socket ignored
Warning: Stopping docker.service, but it can still be activated by:
  docker.socket
tar: var/lib/kubelet/device-plugins/kubelet.sock: socket ignored
* Waiting for kube-proxy and the nodes 
.......[done]
* Bringing up kube-dns and kubernetes-dashboard 
deployment.extensions/kube-dns scaled
deployment.extensions/kubernetes-dashboard scaled
........................[done]
NAME          STATUS    ROLES     AGE       VERSION
kube-master   Ready     master    3m        v1.11.0
kube-node-1   Ready     <none>    2m        v1.11.0
kube-node-2   Ready     <none>    2m        v1.11.0
* Access dashboard at: http://127.0.0.1:32768/api/v1/namespaces/kube-system/services/kubernetes-dashboard:/proxy
customresourcedefinition.apiextensions.k8s.io/podgroups.scheduling.incubator.k8s.io created
error: the path "config/crds/extensions_v1alpha1_job.yaml" does not exist
customresourcedefinition.apiextensions.k8s.io/queues.scheduling.incubator.k8s.io created
=== RUN   TestE2E
Running Suite: kube-arbitrator Test Suite
=========================================
Random Seed: 1536046116
Will run 8 of 8 specs

• Failure [82.406 seconds]
E2E Test
/home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:28
  Schedule Job with SchedulerName [It]
  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:29

  Expected error:
      <*errors.errorString | 0xc4200b7630>: {
          s: "timed out waiting for the condition",
      }
      timed out waiting for the condition
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:35
------------------------------
• Failure [82.603 seconds]
E2E Test
/home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:28
  Schedule Job [It]
  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:38

  Expected error:
      <*errors.errorString | 0xc4200b7630>: {
          s: "timed out waiting for the condition",
      }
      timed out waiting for the condition
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:44
------------------------------
• Failure [66.404 seconds]
E2E Test
/home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:28
  Schedule Multiple Jobs [It]
  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:47

  Expected error:
      <*errors.errorString | 0xc4200b7630>: {
          s: "timed out waiting for the condition",
      }
      timed out waiting for the condition
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:58
------------------------------
• Failure [87.021 seconds]
E2E Test
/home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:28
  Gang scheduling [It]
  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:67

  Expected error:
      <*errors.errorString | 0xc4200b7630>: {
          s: "timed out waiting for the condition",
      }
      timed out waiting for the condition
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:84
------------------------------
• Failure [82.403 seconds]
E2E Test
/home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:28
  Preemption [It]
  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:87

  Expected error:
      <*errors.errorString | 0xc4200b7630>: {
          s: "timed out waiting for the condition",
      }
      timed out waiting for the condition
  not to have occurred

  /home/travis/gopath/src/github.com/kubernetes-incubator/kube-arbitrator/test/e2e.go:96
------------------------------
• Failure [66.203 seconds]
E2E Test
…